requirements.txt
Flask==2.3.3
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
kornia>=0.7.0
flask-cors>=4.0.0
werkzeug>=2.3.7
gunicorn>=21.2.0